Backend Changes

+++ backend/controllers/product.controller.js
@@ -49,11 +49,14 @@ export const updateProduct = async(req, res) => {
 export const deleteProduct = async(req, res) => {
     const { id } = req.params;
     console.log(id);
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Product not found" });
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     } catch (error) {
-        console.error("Product is not deleted", error.message);
-        res.status(500).json({ success: false, message: "Product deletion failed" });
+        console.error("Error in deleting product", error.message);
+        res.status(500).json({ success: false, message: "Server Error", error: error.message });
     }
 }
